Fixed issue with zero-row data.tables
fix: Update in data.table changed an output in the flow calculations, resulting in a zero-row and zero-column data.table instead of a zero-row data.table retaining the original column names and dimensions. This resulted in errors with attempted indexing and merging with the land values. Change simply tests to see if the table is zero-row, and if so, skips to the export.